backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "fastapi[standard]>=0.116.1",
    "google-genai>=1.29.0",
    "passlib[bcrypt]>=1.7.4",
    "bcrypt<4.1",
    "pgvector>=0.4.1",
    "psycopg2-binary>=2.9.10",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.42",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
]
